Virtual ind

See merge request hm171/fantasy!14

+++ src/main/java/edu/duke/ece/fantasy/json/PositionRequestMessage.java
@@ -1,29 +1,29 @@
 package edu.duke.ece.fantasy.json;
 
 public class PositionRequestMessage {
-    private double x;
-    private double y;
+    private int x;
+    private int y;
 
     public PositionRequestMessage(){}
 
-    public PositionRequestMessage(double x, double y) {
+    public PositionRequestMessage(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
 }
